Adding auth to tests

+++ tests/unit/test_party.py
@@ -1,20 +1,20 @@
-def test_add_party(client):
+def test_add_party(client, token):
     response = client.post('/parties', json={
         'name': 'test',
         'votes': 0
-    })
+    }, headers={'Authorization': token})
     assert response.status_code == 201
     assert response.json == {'id': 1, 'name': 'test', 'votes': 0}
 
 
-def test_get_empty_parties(client):
-    response = client.get('/parties')
+def test_get_empty_parties(client, token):
+    response = client.get('/parties', headers={'Authorization': token})
     assert response.status_code == 200
     assert response.json == []
 
 
-def test_get_parties(client, populate_party):
-    response = client.get('/parties')
+def test_get_parties(client, populate_party, token):
+    response = client.get('/parties', headers={'Authorization': token})
     assert response.status_code == 200
     assert response.json == [
         {'id': 1, 'name': 'partyA', 'votes': 340000},
@@ -25,39 +25,39 @@ def test_get_parties(client, populate_party):
     ]
 
 
-def test_get_empty_party(client):
-    response = client.get('/parties/1')
+def test_get_empty_party(client, token):
+    response = client.get('/parties/1', headers={'Authorization': token})
     assert response.status_code == 404
 
 
-def test_get_party(client, populate_party):
-    response = client.get('/parties/1')
+def test_get_party(client, populate_party, token):
+    response = client.get('/parties/1', headers={'Authorization': token})
     assert response.status_code == 200
     assert response.json == {'id': 1, 'name': 'partyA', 'votes': 340000}
 
 
-def test_edit_empty_party(client):
+def test_edit_empty_party(client, token):
     response = client.put('/parties/1', json={
         'name': 'test',
         'votes': 0
-    })
+    }, headers={'Authorization': token})
     assert response.status_code == 404
 
 
-def test_edit_party(client, populate_party):
+def test_edit_party(client, populate_party, token):
     response = client.put('/parties/1', json={
         'name': 'test',
         'votes': 0
-    })
+    }, headers={'Authorization': token})
     assert response.status_code == 201
     assert response.json == {'id': 1, 'name': 'test', 'votes': 0}
 
 
-def test_delete_empty_party(client):
-    response = client.delete('/parties/1')
+def test_delete_empty_party(client, token):
+    response = client.delete('/parties/1', headers={'Authorization': token})
     assert response.status_code == 404
 
 
-def test_delete_party(client, populate_party):
-    response = client.delete('/parties/1')
+def test_delete_party(client, populate_party, token):
+    response = client.delete('/parties/1', headers={'Authorization': token})
     assert response.status_code == 204
